+++ Assets/CommonCore/Core/CoreParams.cs
@@ -67,6 +67,7 @@ namespace CommonCore
         public static bool AllowSaveLoad { get; private set; } = true;
         public static bool AllowManualSave { get; private set; } = true;
         public static bool EnforceQuicksaveVersionMatching { get; private set; } = true;
+        public static bool KeepAutosaveForEachCampaign { get; private set; } = true;
         public static IReadOnlyList<string> AdditionalAxes { get; private set; } = ImmutableArray.Create<string>(); //specify additional axes your game will use; it's up to individual input mappers to handle these
         public static IReadOnlyList<string> AdditionalButtons { get; private set; } = ImmutableArray.Create<string>(); //same, but for buttons
         public static IReadOnlyList<string> HideControlMappings { get; private set; } = ImmutableArray.Create<string>("OpenMenu"); //add things to this to hide DefaultControls you're not using, note that it's not guaranteed to stop the control from responding to input